routines to set up distributed environments are moved to a separate file from utils.py
