+++ utils/__init__.py
@@ -6,5 +6,6 @@ from .utils import (
     train_validate_test_normal,
     test,
 )
+from .distributed_environment import get_comm_size_and_rank
 from .print_utils import print_distributed, iterate_tqdm
 from .visualizer import Visualizer
