+++ utils/utils.py
@@ -21,118 +21,11 @@ from data_utils.dataset_descriptors import (
     StructureFeatures,
     Dataset,
 )
+from utils.distributed_environment import get_comm_size_and_rank
 from utils.print_utils import print_distributed, iterate_tqdm
 from utils.visualizer import Visualizer
 
 import re
-
-
-def parse_slurm_nodelist(nodelist):
-    """
-    Parse SLURM_NODELIST env string to get list of nodes.
-    Usage example:
-        parse_slurm_nodelist(os.environ["SLURM_NODELIST"])
-    Input examples:
-        "or-condo-g04"
-        "or-condo-g[05,07-08,13]"
-        "or-condo-g[05,07-08,13],or-condo-h[01,12]"
-    """
-    nlist = list()
-    for block, _ in re.findall(r"([\w-]+(\[[\d\-,]+\])*)", nodelist):
-        m = re.match(r"^(?P<prefix>[\w\-]+)\[(?P<group>.*)\]", block)
-        if m is None:
-            ## single node
-            nlist.append(block)
-        else:
-            ## multiple nodes
-            g = m.groups()
-            prefix = g[0]
-            for sub in g[1].split(","):
-                if "-" in sub:
-                    start, end = re.match(r"(\d+)-(\d+)", sub).groups()
-                    fmt = "%%0%dd" % (len(start))
-                    for i in range(int(start), int(end) + 1):
-                        node = prefix + fmt % i
-                        nlist.append(node)
-                else:
-                    node = prefix + sub
-                    nlist.append(node)
-
-    return nlist
-
-
-def init_comm_size_and_rank():
-    world_size = None
-    world_rank = 0
-
-    if os.getenv("OMPI_COMM_WORLD_SIZE") and os.getenv("OMPI_COMM_WORLD_RANK"):
-        ## Summit
-        world_size = int(os.environ["OMPI_COMM_WORLD_SIZE"])
-        world_rank = int(os.environ["OMPI_COMM_WORLD_RANK"])
-    elif os.getenv("SLURM_NPROCS") and os.getenv("SLURM_PROCID"):
-        ## CADES
-        world_size = int(os.environ["SLURM_NPROCS"])
-        world_rank = int(os.environ["SLURM_PROCID"])
-
-    ## Fall back to default
-    if world_size is None:
-        world_size = 1
-
-    return int(world_size), int(world_rank)
-
-
-def get_comm_size_and_rank():
-    world_size = None
-    world_rank = 0
-
-    if dist.is_initialized():
-        world_size = dist.get_world_size()
-        world_rank = dist.get_rank()
-    else:
-        world_size = 1
-
-    return int(world_size), int(world_rank)
-
-
-def setup_ddp():
-
-    """ "Initialize DDP"""
-
-    if dist.is_nccl_available() and torch.cuda.is_available():
-        backend = "nccl"
-    elif torch.distributed.is_gloo_available():
-        backend = "gloo"
-    else:
-        raise RuntimeError("No parallel backends available")
-
-    world_size, world_rank = init_comm_size_and_rank()
-
-    ## Default setting
-    master_addr = "127.0.0.1"
-    master_port = "8889"
-
-    if os.getenv("LSB_HOSTS") is not None:
-        ## source: https://www.olcf.ornl.gov/wp-content/uploads/2019/12/Scaling-DL-on-Summit.pdf
-        ## The following is Summit specific
-        master_addr = os.environ["LSB_HOSTS"].split()[1]
-    elif os.getenv("SLURM_NODELIST") is not None:
-        ## The following is CADES specific
-        master_addr = parse_slurm_nodelist(os.environ["SLURM_NODELIST"])[0]
-
-    try:
-        os.environ["MASTER_ADDR"] = master_addr
-        os.environ["MASTER_PORT"] = master_port
-        os.environ["WORLD_SIZE"] = str(world_size)
-        os.environ["RANK"] = str(world_rank)
-        if not dist.is_initialized():
-            dist.init_process_group(
-                backend=backend, rank=int(world_rank), world_size=int(world_size)
-            )
-    except KeyError:
-        print("DDP has to be initialized within a job - Running in sequential mode")
-
-    return world_size, world_rank
-
 
 def train_validate_test_normal(
     model,
